Refactor : 그룹 선택 코드 수정
