+++ src/App.tsx
@@ -4,8 +4,8 @@ import { ApolloProvider } from "@apollo/client";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { client } from "../src/apollo/client";
 import Layout from "./components/layout";
-import SelectGroupList from "./components/pages/SelectGroup";
-import SearchCafe from "./components/pages/SearchCafe";
+import SelectGroupList from "./pages/SelectGroup";
+import SearchCafe from "./pages/SearchCafe";
 import "./App.css";
 
 function App() {
