+++ src/apollo/gql/selectGroupList.gql.tsx
@@ -1,0 +1,18 @@
+import { gql } from "@apollo/client";
+
+const GROUP_USER = gql`
+query {
+    groups {
+      _id
+      name
+      debutDate
+      logo{
+        name
+        filename
+        _id
+      }
+    }
+  }
+`;
+
+export default GROUP_USER;
